Added new monster. Simple attack code added back to player.
Former-commit-id: 453001755b4f625d60b871e7bedb3697b33dc89b

+++ src/gmap.cpp
@@ -34,7 +34,7 @@ inline u64 HashStringJenkins(u64 Size, const void* Key)
 }
 
 inline u64 HashInt(u64 Size, const void* Key) {
-    u64 K = *(u64*)(Key);
+    u64 K = (u64)(*(i32*)(Key));
     K = ((K >> 16) ^ K) * 0x45d9f3b;
     K = ((K >> 16) ^ K) * 0x45d9f3b;
     K = (K >> 16) ^ K;
